--- a/floats/floats.go
+++ b/floats/floats.go
@@ -7,6 +7,7 @@ package floats
 import (
 	"errors"
 	"math"
+	"slices"
 	"sort"
 
 	"gonum.org/v1/gonum/floats/scalar"
@@ -634,9 +635,7 @@ func Prod(s []float64) float64 {
 
 // Reverse reverses the order of elements in the slice.
 func Reverse(s []float64) {
-	for i, j := 0, len(s)-1; i < j; i, j = i+1, j-1 {
-		s[i], s[j] = s[j], s[i]
-	}
+	slices.Reverse(s)
 }
 
 // Same returns true when the input slices have the same length and all