--- a/floats/floats.go
+++ b/floats/floats.go
@@ -633,6 +633,8 @@ func Prod(s []float64) float64 {
 	return prod
 }
 
+// Deprecated: This function simply calls [slices.Reverse].
+//
 // Reverse reverses the order of elements in the slice.
 func Reverse(s []float64) {
 	slices.Reverse(s)