@jbduncan 2024-03-15
@kortschak Running the benchmarks in [ced1a5f](https://github.com/gonum/gonum/pull/1944/commits/ced1a5f9a91ca4d78fe4e0f21347796e17a1560e), the results look pretty much the same. What do you think?

```
goos: darwin
goarch: arm64
pkg: gonum.org/v1/gonum/floats
                │ benchmarks_floats_reverse_old.txt │ benchmarks_floats_reverse_new.txt  │
                │              sec/op               │   sec/op     vs base               │
ReverseSmall-8                          2.843n ± 1%   2.817n ± 0%  -0.90% (p=0.006 n=10)
ReverseMedium-8                         241.7n ± 1%   240.3n ± 0%  -0.56% (p=0.040 n=10)
ReverseLarge-8                          23.76µ ± 1%   23.57µ ± 1%       ~ (p=0.165 n=10)
ReverseHuge-8                           2.822m ± 0%   2.805m ± 1%       ~ (p=0.052 n=10)
geomean                                 2.605µ        2.586µ       -0.71%
```